modules/filterModule/requirements.txt
azure-iot-device>=2.5.0